Made resources a module, cleaned up
