+++ character.py
@@ -1,28 +1,30 @@
 import resources
 
-
 # Requirements for player are an existing control scheme and
-# 1 preview and 6 running frames as standard images stored in resources/img/character/{character}
+# 1 preview and 6 running frames as standard images stored in resources/images/character/{character}
+from resources.schemes import Scheme
 
 
 class Character:
 
-    def __init__(self, texture: str, control_scheme: str, name: str):
+    def __init__(self, texture: str, scheme: Scheme, name: str):
         self.texture = texture
         self.name = name
 
         # preview image for the character
         path = "character/" + texture
-        self.preview_img = resources.load_img(path + "_preview.png")
+        self.preview_img = resources.images.load_img(path + "_preview.png")
 
         # immutable collection of run animation frames consisting of 6 frames
         path = path + "_run_f{}.png"
         self.run_anim = (
-            resources.load_img(path.format(0)), resources.load_img(path.format(1)), resources.load_img(path.format(2)),
-            resources.load_img(path.format(3)), resources.load_img(path.format(4)), resources.load_img(path.format(5)))
+            resources.images.load_img(path.format(0)), resources.images.load_img(path.format(1)),
+            resources.images.load_img(path.format(2)),
+            resources.images.load_img(path.format(3)), resources.images.load_img(path.format(4)),
+            resources.images.load_img(path.format(5)))
         self.cur_frame = 0
         # loading the control scheme
-        self.scheme = Scheme(control_scheme)
+        self.scheme = scheme
         # setting game related parameters
         self.health = 3
         self.can_exit = False
@@ -34,16 +36,3 @@ class Character:
         if self.health >= 0:
             self.health = 0
             self.dead = True
-
-
-class Scheme:
-
-    def __init__(self, control_scheme: str):
-        controls = resources.load_scheme(control_scheme)
-
-        self.up = controls[0]
-        self.down = controls[1]
-        self.left = controls[2]
-        self.right = controls[3]
-        self.action = controls[4]
-        self.switch = controls[5]
