--- game.py
@@ -1,3 +1,0 @@
-import pygame
-
-# placeholder for game loop script
