+++ resources/__init__.py
@@ -1,0 +1,6 @@
+"""
+@author: daniel.fedotov
+"""
+from resources import images
+from resources import levels
+from resources import schemes
