+++ resources/__resources_test.py
@@ -1,20 +1,21 @@
 import unittest
-import resources
-import pygame
+import images
+import levels
+import schemes
 
 
 class MyTestCase(unittest.TestCase):
 
     def test_load_img_gob_preview(self):
-        test_img = resources.load_img("character/goblin/_preview.png")
+        test_img = images.load_img("character/goblin/_preview.png")
         self.assertEqual((16, 16), test_img.get_size())
 
     def test_load_img_dir_floor(self):
-        floor_images = resources.load_img_dir("tex/floor")
+        floor_images = images.load_directory("tex/floor")
         print(floor_images)
 
     def test_load_scheme_wasd(self):
-        test_scheme_wasd = resources.load_scheme("wasd")
+        test_scheme_wasd = schemes.load_scheme("wasd")
         self.assertEqual(119, test_scheme_wasd[0])
         self.assertEqual(115, test_scheme_wasd[1])
         self.assertEqual(97, test_scheme_wasd[2])
@@ -23,7 +24,7 @@ class MyTestCase(unittest.TestCase):
         self.assertEqual(113, test_scheme_wasd[5])
 
     def test_load_prebuilt(self):
-        test_pattern = resources.load_prebuilt_pattern("maze_test")
+        test_pattern = levels.load_pattern("maze_test")
         print(test_pattern)
         self.assertEqual(["f", "end"], test_pattern[0][0])
         self.assertEqual(["f"], test_pattern[1][0])
@@ -32,7 +33,7 @@ class MyTestCase(unittest.TestCase):
         self.assertEqual(["f"], test_pattern[9][9])
 
     def test_load_broken_prebuilt(self):
-        test_pattern = resources.load_prebuilt_pattern("test/broken")
+        test_pattern = levels.load_pattern("test/broken")
         print(test_pattern)
         self.assertEqual(["f", "f"], test_pattern[0][0])
         self.assertEqual(["f", "f", "w"], test_pattern[0][7])
