+++ resources/images/character/goblin/_preview.png
