+++ resources/images/character/goblin/_run_f2.png
