+++ resources/images/tex/floor/weight.txt
