+++ resources/images/tex/wall/weight.txt
