+++ resources/images/ui/exit/hover.png
