+++ resources/images/ui/exit/pressed.png
