+++ tilescript.py
@@ -1,4 +1,4 @@
-import resources
+from resources import _resources
 
 
 def __effect_key(maze, character, pos):
@@ -35,10 +35,10 @@ def __effect_on_set_end(maze, character, pos):
 # [4] is this a base tile? Base tiles will be ignored if attempt is made to load them after the first layer
 # [5] is this tile unique? When second instance of this tile is encountered it will be ignored.
 dictionary = {
-    'f': (resources.load_img_dir("tex/floor"), 0, None, None, True, False),
+    'f': (_resources.load_directory("tex/floor"), 0, None, None, True, False),
     'start': (None, 0, None, __effect_on_set_player, False, True),
-    'w': (resources.load_img_dir("tex/wall"), 999, None, None, True, False),
-    'key': ((resources.load_img("tex/key.png"),), -3, __effect_key, __effect_on_set_key, False, True),
-    'end': ((resources.load_img("tex/door.png"),), 0, __effect_win, __effect_on_set_end, False, True),
-    'trap': ((resources.load_img("tex/trap.png"),), 2, __effect_damage_light, None, True, False)
+    'w': (_resources.load_directory("tex/wall"), 999, None, None, True, False),
+    'key': ((_resources.load_img("tex/key.png"),), -3, __effect_key, __effect_on_set_key, False, True),
+    'end': ((_resources.load_img("tex/door.png"),), 0, __effect_win, __effect_on_set_end, False, True),
+    'trap': ((_resources.load_img("tex/trap.png"),), 2, __effect_damage_light, None, True, False)
 }
